backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
passlib
python-jose
pymysql
Pillow
python-multipart
